Refactor createStandardServiceItem template
