+++ src/constants/MoveHistory/EventTemplates/ApproveShipment/createStandardServiceItem.jsx
@@ -1,5 +1,6 @@
+import React from 'react';
+
 import o from 'constants/MoveHistory/UIDisplay/Operations';
-import d from 'constants/MoveHistory/UIDisplay/DetailsTypes';
 import a from 'constants/MoveHistory/Database/Actions';
 import t from 'constants/MoveHistory/Database/Tables';
 import { shipmentTypes as SHIPMENT_TYPE } from 'constants/shipments';
@@ -9,12 +10,14 @@ export default {
   action: a.INSERT,
   eventName: o.approveShipment,
   tableName: t.mto_service_items,
-  detailsType: d.PLAIN_TEXT,
   getEventNameDisplay: () => 'Approved service item',
-  getDetailsPlainText: (historyRecord) => {
+  getDetails: (historyRecord) => {
     const shipmentLabel = getMtoShipmentLabel(historyRecord);
-    return `${SHIPMENT_TYPE[shipmentLabel.shipment_type]} shipment #${shipmentLabel.shipment_id_display}, ${
-      shipmentLabel.service_item_name
-    }`;
+    return (
+      <>
+        {SHIPMENT_TYPE[shipmentLabel.shipment_type]} shipment #{shipmentLabel.shipment_id_display},
+        {shipmentLabel.service_item_name}
+      </>
+    );
   },
 };
