+++ src/constants/MoveHistory/EventTemplates/ApproveShipment/createStandardServiceItem.test.jsx
@@ -1,8 +1,9 @@
+import { render, screen } from '@testing-library/react';
+
 import getTemplate from 'constants/MoveHistory/TemplateManager';
-import e from 'constants/MoveHistory/EventTemplates/createStandardServiceItem';
 
 describe('when given a Create standard service item history record', () => {
-  const item = {
+  const historyRecord = {
     action: 'INSERT',
     context: [
       {
@@ -15,9 +16,10 @@ describe('when given a Create standard service item history record', () => {
     tableName: 'mto_service_items',
   };
   it('correctly matches the Create standard service item event', () => {
-    const result = getTemplate(item);
-    expect(result).toMatchObject(e);
-    expect(result.getEventNameDisplay(result)).toEqual('Approved service item');
-    expect(result.getDetailsPlainText(item)).toEqual('HHG shipment #A1B2C, Domestic linehaul');
+    const template = getTemplate(historyRecord);
+    render(template.getDetails(historyRecord));
+    expect(template.getEventNameDisplay(template)).toEqual('Approved service item');
+    expect(screen.getByText('HHG shipment #A1B2C', { exact: false })).toBeInTheDocument();
+    expect(screen.getByText('Domestic linehaul', { exact: false })).toBeInTheDocument();
   });
 });
